# api/openapi.yaml
openapi: 3.1.0
info:
    title: 'Voting API'
    version: 1.0.0
    description: 'Secure voting system API — Express, TypeScript, Prisma. Auto-generated from JSDoc comments.'
servers:
    -
        url: 'http://localhost:4000'
        description: 'Local development'
    -
        url: 'https://api.yourdomain.com'
        description: Production
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: 'Paste an **access token** here.'
security:
    -
        bearerAuth: []
...
paths:
...
    '/api/candidates/position/{positionId}':
        get:
            tags:
                - Candidates
            summary: 'Get candidates by position'
            description: 'Retrieve all candidates for a specific position'
            security:
                -
                    bearerAuth: []
            parameters:
                -
                    name: positionId
                    in: path
                    required: true
                    description: 'Position ID'
                    schema:
                        type: string
                        format: uuid
            responses:
                '200':
                    description: 'Candidates retrieved successfully'
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: object
                                    properties: {id: {type: string, format: uuid}, fullName: {type: string}, email: {type: string, format: email}, phone: {type: string}, bio: {type: string}, manifesto: {type: string}, order: {type: number}, isActive: {type: boolean}, position: {type: object, properties: {id: {type: string, format: uuid}, name: {type: string}, election: {type: object, properties: {id: {type: string, format: uuid}, title: {type: string}, status: {type: string}}}}}, _count: {type: object, properties: {votes: {type: number}}}}
                '401':
                    description: 'Unauthorized - Authentication required'
                '403':
                    description: 'Forbidden - EC access required'
    '/api/candidates/election/{electionId}':
        get:
            tags:
                - Candidates
            summary: 'Get candidates by election'
            description: 'Retrieve all candidates for a specific election'
            security:
                -
                    bearerAuth: []
            parameters:
                -
                    name: electionId
                    in: path
                    required: true
                    description: 'Election ID'
                    schema:
                        type: string
                        format: uuid
            responses:
                '200':
                    description: 'Candidates retrieved successfully'
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: object
                                    properties: {id: {type: string, format: uuid}, fullName: {type: string}, email: {type: string, format: email}, phone: {type: string}, bio: {type: string}, manifesto: {type: string}, order: {type: number}, isActive: {type: boolean}, position: {type: object, properties: {id: {type: string, format: uuid}, name: {type: string}, order: {type: number}}}, _count: {type: object, properties: {votes: {type: number}}}}
                '401':
                    description: 'Unauthorized - Authentication required'
                '403':
                    description: 'Forbidden - EC access required'
...